feat(Lotto): 6개의 숫자를 가지고 있는 로또 기능
Lotto 클래스는 다음과 같은 기능을 가지고 있다.
- 숫자의 개수가 6개인지 확인
- 중복되는 숫자가 없는지 확인
- 6개의 숫자를 오름차순으로 정렬해 문자열로 반환하는 기능
- 숫자가 리스트에 있는지 확인하는 기능
