+++ src/main/java/lotto/model/Lotto.java
@@ -1,0 +1,44 @@
+package lotto.model;
+
+import static lotto.constant.ErrorMessage.DUPLICATED_LOTTO_NUMBER;
+import static lotto.constant.ErrorMessage.INVALID_LOTTO_NUMBER_COUNT;
+import static lotto.constant.LottoInfo.LOTTO_NUMBER_COUNT;
+import static lotto.constant.LottoInfo.MAX_LOTTO_NUMBER;
+import static lotto.constant.LottoInfo.MIN_LOTTO_NUMBER;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import java.util.List;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    public Lotto() {
+        this.numbers = Randoms.pickUniqueNumbersInRange(MIN_LOTTO_NUMBER, MAX_LOTTO_NUMBER, LOTTO_NUMBER_COUNT);
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != LOTTO_NUMBER_COUNT) {
+            throw new IllegalArgumentException(INVALID_LOTTO_NUMBER_COUNT.getMessage());
+        }
+
+        if (numbers.stream().distinct().count() != LOTTO_NUMBER_COUNT) {
+            throw new IllegalArgumentException(DUPLICATED_LOTTO_NUMBER.getMessage());
+        }
+    }
+
+    public List<String> toSortedStrings() {
+        return numbers.stream()
+                .sorted()
+                .map(number -> number + "")
+                .toList();
+    }
+
+    public boolean contains(int number) {
+        return numbers.contains(number);
+    }
+}
